plat/nxp: lx2160a auto boot
This patch adds support to patch RCW that already has SD/eMMC/SPI boot
support embedded with conditional load and jump.
The idea is to look for SD/eMMC/SPI boot, and modify src/dst/size
address with the correct values; rather than adding blockread at the end
of RCW code.

With this patch images are unified and can be used to boot from SD /
eMMC and SPI.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>
Signed-off-by: Grant Likely <grant.likely@arm.com>